Complex Number Matrix Multiplier
Final project for my C class: it reads two matrices of complex numbers from a file and displays their product. It's not user-friendly since it was schoolwork, but I'm planning on making it more finished and more functional. (Likely, this will take some time to get around to.)

See the ReadMe.txt for more information.